+++ src/test/java/com/iexec/resultproxy/authorization/AuthorizationServiceTests.java
@@ -21,11 +21,9 @@ import com.iexec.commons.poco.chain.ChainDeal;
 import com.iexec.commons.poco.chain.ChainTask;
 import com.iexec.commons.poco.chain.WorkerpoolAuthorization;
 import com.iexec.commons.poco.security.Signature;
-import com.iexec.commons.poco.tee.TeeUtils;
 import com.iexec.commons.poco.utils.BytesUtils;
 import com.iexec.commons.poco.utils.HashUtils;
 import com.iexec.commons.poco.utils.SignatureUtils;
-import com.iexec.commons.poco.utils.TestUtils;
 import com.iexec.resultproxy.chain.IexecHubService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -48,11 +46,13 @@ import java.security.GeneralSecurityException;
 import java.security.InvalidAlgorithmParameterException;
 import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
+import java.time.Instant;
+import java.time.temporal.ChronoUnit;
 import java.util.Optional;
 
 import static com.iexec.commons.poco.chain.ChainTaskStatus.ACTIVE;
-import static com.iexec.commons.poco.chain.ChainTaskStatus.UNSET;
 import static com.iexec.commons.poco.utils.SignatureUtils.signMessageHashAndGetSignature;
+import static com.iexec.resultproxy.TestUtils.*;
 import static com.iexec.resultproxy.authorization.AuthorizationError.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
@@ -61,9 +61,7 @@ import static org.mockito.Mockito.when;
 @Testcontainers
 class AuthorizationServiceTests {
 
-    private static final String CHAIN_TASK_ID = "0x0123";
     private static final String RESULT_DIGEST = "0x3210";
-    private static final String WALLET_ADDRESS = "0xabcd";
 
     @Container
     private static final MongoDBContainer mongoDBContainer = new MongoDBContainer(DockerImageName.parse(System.getProperty("mongo.image")));
@@ -82,11 +80,13 @@ class AuthorizationServiceTests {
     private AuthorizationService authorizationService;
 
     private Credentials enclaveCreds;
+    private Credentials workerCreds;
 
     @BeforeEach
     void beforeEach() throws GeneralSecurityException {
         MockitoAnnotations.openMocks(this);
         enclaveCreds = Credentials.create(Keys.createEcKeyPair());
+        workerCreds = Credentials.create(Keys.createEcKeyPair());
         authorizationRepository.deleteAll();
         authorizationService = new AuthorizationService(authorizationRepository, iexecHubService);
     }
@@ -94,9 +94,9 @@ class AuthorizationServiceTests {
     // region isAuthorizedOnExecutionWithDetailedIssue
     @Test
     void shouldBeAuthorizedOnExecutionOfTeeTaskWithDetails() {
-        ChainDeal chainDeal = getChainDeal();
-        ChainTask chainTask = TestUtils.getChainTask(ACTIVE);
-        WorkerpoolAuthorization auth = TestUtils.getTeeWorkerpoolAuth();
+        final ChainDeal chainDeal = getChainDeal();
+        final ChainTask chainTask = getChainTask(ACTIVE);
+        final WorkerpoolAuthorization auth = getWorkerpoolAuthorization(true);
         when(iexecHubService.getChainTask(auth.getChainTaskId())).thenReturn(Optional.of(chainTask));
         when(iexecHubService.getChainDeal(chainTask.getDealid())).thenReturn(Optional.of(chainDeal));
 
@@ -107,106 +107,96 @@ class AuthorizationServiceTests {
     @Test
     void shouldNotBeAuthorizedOnExecutionOfTeeTaskWithNullAuthorizationWithDetails() {
         Optional<AuthorizationError> isAuth = authorizationService.isAuthorizedOnExecutionWithDetailedIssue(null);
-        assertThat(isAuth)
-                .isNotEmpty()
-                .contains(EMPTY_PARAMS_UNAUTHORIZED);
+        assertThat(isAuth).isEqualTo(Optional.of(EMPTY_PARAMS_UNAUTHORIZED));
     }
 
     @Test
     void shouldNotBeAuthorizedOnExecutionOfTeeTaskWithEmptyAuthorizationWithDetails() {
         Optional<AuthorizationError> isAuth = authorizationService.isAuthorizedOnExecutionWithDetailedIssue(WorkerpoolAuthorization.builder().build());
-        assertThat(isAuth).isNotEmpty()
-                .contains(EMPTY_PARAMS_UNAUTHORIZED);
+        assertThat(isAuth).isEqualTo(Optional.of(EMPTY_PARAMS_UNAUTHORIZED));
     }
 
     @Test
     void shouldNotBeAuthorizedOnExecutionOfTeeTaskWhenTaskTypeNotMatchedOnchainWithDetails() {
-        ChainDeal chainDeal = getChainDeal();
-        ChainTask chainTask = TestUtils.getChainTask(ACTIVE);
-        WorkerpoolAuthorization auth = WorkerpoolAuthorization.builder()
+        final ChainDeal chainDeal = getChainDeal();
+        final ChainTask chainTask = getChainTask(ACTIVE);
+        final WorkerpoolAuthorization auth = WorkerpoolAuthorization.builder()
                 .chainTaskId("0x1111111111111111111111111111111111111111111111111111111111111111")
                 .workerWallet("0x87ae2b87b5db23830572988fb1f51242fbc471ce")
-                .enclaveChallenge(BytesUtils.EMPTY_HEX_STRING_32)
+                .enclaveChallenge(BytesUtils.EMPTY_ADDRESS)
                 .build();
         when(iexecHubService.getChainTask(auth.getChainTaskId())).thenReturn(Optional.of(chainTask));
         when(iexecHubService.getChainDeal(chainTask.getDealid())).thenReturn(Optional.of(chainDeal));
 
         Optional<AuthorizationError> isAuth = authorizationService.isAuthorizedOnExecutionWithDetailedIssue(auth);
-        assertThat(isAuth)
-                .isNotEmpty()
-                .contains(NO_MATCH_ONCHAIN_TYPE);
+        assertThat(isAuth).isEqualTo(Optional.of(NO_MATCH_ONCHAIN_TYPE));
     }
 
     @Test
     void shouldNotBeAuthorizedOnExecutionOfTeeTaskWhenGetTaskFailedWithDetails() {
-        WorkerpoolAuthorization auth = TestUtils.getTeeWorkerpoolAuth();
+        final WorkerpoolAuthorization auth = getWorkerpoolAuthorization(true);
         when(iexecHubService.isTeeTask(auth.getChainTaskId())).thenReturn(true);
         when(iexecHubService.getChainTask(auth.getChainTaskId())).thenReturn(Optional.empty());
 
         Optional<AuthorizationError> isAuth = authorizationService.isAuthorizedOnExecutionWithDetailedIssue(auth);
-        assertThat(isAuth)
-                .isNotEmpty()
-                .contains(GET_CHAIN_TASK_FAILED);
+        assertThat(isAuth).isEqualTo(Optional.of(GET_CHAIN_TASK_FAILED));
     }
 
     @Test
-    void shouldNotBeAuthorizedOnExecutionOfTeeTaskWhenTaskNotActiveWithDetails() {
-        WorkerpoolAuthorization auth = TestUtils.getTeeWorkerpoolAuth();
-        ChainTask chainTask = TestUtils.getChainTask(UNSET);
+    void shouldNotBeAuthorizedOnExecutionOfTeeTaskWhenFinalDeadlineReached() {
+        final WorkerpoolAuthorization auth = getWorkerpoolAuthorization(true);
+        final ChainTask chainTask = ChainTask.builder()
+                .dealid(CHAIN_DEAL_ID)
+                .finalDeadline(Instant.now().minus(5L, ChronoUnit.SECONDS).toEpochMilli())
+                .build();
         when(iexecHubService.isTeeTask(auth.getChainTaskId())).thenReturn(true);
         when(iexecHubService.getChainTask(auth.getChainTaskId())).thenReturn(Optional.of(chainTask));
 
         Optional<AuthorizationError> isAuth = authorizationService.isAuthorizedOnExecutionWithDetailedIssue(auth);
-        assertThat(isAuth)
-                .isNotEmpty()
-                .contains(TASK_NOT_ACTIVE);
+        assertThat(isAuth).isEqualTo(Optional.of(TASK_FINAL_DEADLINE_REACHED));
     }
 
     @Test
     void shouldNotBeAuthorizedOnExecutionOfTeeTaskWhenGetDealFailedWithDetails() {
-        ChainTask chainTask = TestUtils.getChainTask(ACTIVE);
-        WorkerpoolAuthorization auth = TestUtils.getTeeWorkerpoolAuth();
-        auth.setSignature(new Signature(TestUtils.POOL_WRONG_SIGNATURE));
+        final ChainTask chainTask = getChainTask(ACTIVE);
+        final WorkerpoolAuthorization auth = getWorkerpoolAuthorization(true);
+        auth.setSignature(new Signature(POOL_WRONG_SIGNATURE));
 
         when(iexecHubService.isTeeTask(auth.getChainTaskId())).thenReturn(true);
         when(iexecHubService.getChainTask(auth.getChainTaskId())).thenReturn(Optional.of(chainTask));
         when(iexecHubService.getChainDeal(chainTask.getDealid())).thenReturn(Optional.empty());
 
         Optional<AuthorizationError> isAuth = authorizationService.isAuthorizedOnExecutionWithDetailedIssue(auth);
-        assertThat(isAuth)
-                .isNotEmpty()
-                .contains(GET_CHAIN_DEAL_FAILED);
+        assertThat(isAuth).isEqualTo(Optional.of(GET_CHAIN_DEAL_FAILED));
     }
 
     @Test
     void shouldNotBeAuthorizedOnExecutionOfTeeTaskWhenPoolSignatureIsNotValidWithDetails() {
-        ChainDeal chainDeal = getChainDeal();
-        ChainTask chainTask = TestUtils.getChainTask(ACTIVE);
-        WorkerpoolAuthorization auth = TestUtils.getTeeWorkerpoolAuth();
-        auth.setSignature(new Signature(TestUtils.POOL_WRONG_SIGNATURE));
+        final ChainDeal chainDeal = getChainDeal();
+        final ChainTask chainTask = getChainTask(ACTIVE);
+        final WorkerpoolAuthorization auth = getWorkerpoolAuthorization(true);
+        auth.setSignature(new Signature(POOL_WRONG_SIGNATURE));
 
         when(iexecHubService.getChainTask(auth.getChainTaskId())).thenReturn(Optional.of(chainTask));
         when(iexecHubService.getChainDeal(chainTask.getDealid())).thenReturn(Optional.of(chainDeal));
 
         Optional<AuthorizationError> isAuth = authorizationService.isAuthorizedOnExecutionWithDetailedIssue(auth);
-        assertThat(isAuth)
-                .isNotEmpty()
-                .contains(INVALID_SIGNATURE);
+        assertThat(isAuth).isEqualTo(Optional.of(INVALID_SIGNATURE));
     }
     // endregion
 
     @Test
     void checkIsSignedByHimself() throws InvalidAlgorithmParameterException, NoSuchAlgorithmException, NoSuchProviderException {
-        ECKeyPair ecKeyPair = Keys.createEcKeyPair();
-        String privateKey = Numeric.toHexStringWithPrefix(ecKeyPair.getPrivateKey());
-        String workerWallet = Credentials.create(ecKeyPair).getAddress();
-        WorkerpoolAuthorization authorization = WorkerpoolAuthorization.builder()
+        final ECKeyPair ecKeyPair = Keys.createEcKeyPair();
+        final String privateKey = Numeric.toHexStringWithPrefix(ecKeyPair.getPrivateKey());
+        final String workerWallet = Credentials.create(ecKeyPair).getAddress();
+        final WorkerpoolAuthorization authorization = WorkerpoolAuthorization.builder()
                 .chainTaskId("0x1234")
                 .enclaveChallenge("0x5678")
                 .workerWallet(workerWallet)
                 .build();
-        String challenge = authorizationService.getChallengeForWorker(authorization);
-        String signedChallenge = signMessageHashAndGetSignature(challenge, privateKey).getValue();
+        final String challenge = authorizationService.getChallengeForWorker(authorization);
+        final String signedChallenge = signMessageHashAndGetSignature(challenge, privateKey).getValue();
         assertThat(authorizationService.isSignedByHimself(challenge, signedChallenge, workerWallet)).isTrue();
         assertThat(authorizationService.isSignedByHimself(challenge, signedChallenge, Keys.toChecksumAddress(workerWallet))).isTrue();
     }
@@ -225,14 +215,14 @@ class AuthorizationServiceTests {
     // region workerpool authorization cache
     @Test
     void shouldNotBeSignedByEnclaveWhenEnclaveSignatureIsEmpty() {
-        final WorkerpoolAuthorization authorization = getWorkerpoolAuthorization();
+        final WorkerpoolAuthorization authorization = getWorkerpoolAuthorization(true);
         authorizationService.putIfAbsent(authorization);
         final ResultModel model = ResultModel.builder()
                 .chainTaskId(CHAIN_TASK_ID)
                 .enclaveSignature(ResultModel.EMPTY_WEB3_SIG)
                 .deterministHash(RESULT_DIGEST)
                 .build();
-        assertThat(authorizationService.checkEnclaveSignature(model, WALLET_ADDRESS)).isFalse();
+        assertThat(authorizationService.checkEnclaveSignature(model, workerCreds.getAddress())).isFalse();
     }
 
     @Test
@@ -242,66 +232,69 @@ class AuthorizationServiceTests {
                 .enclaveSignature(getEnclaveSignature(enclaveCreds.getEcKeyPair()))
                 .deterministHash(RESULT_DIGEST)
                 .build();
-        assertThat(authorizationService.checkEnclaveSignature(model, WALLET_ADDRESS)).isFalse();
+        assertThat(authorizationService.checkEnclaveSignature(model, workerCreds.getAddress())).isFalse();
     }
 
     @Test
     void shouldNotBeSignedByEnclaveWhenSignedByOther() throws GeneralSecurityException {
-        final WorkerpoolAuthorization authorization = getWorkerpoolAuthorization();
+        final WorkerpoolAuthorization authorization = getWorkerpoolAuthorization(true);
         authorizationService.putIfAbsent(authorization);
         final ResultModel model = ResultModel.builder()
                 .chainTaskId(CHAIN_TASK_ID)
                 .enclaveSignature(getEnclaveSignature(Keys.createEcKeyPair()))
                 .deterministHash(RESULT_DIGEST)
                 .build();
-        assertThat(authorizationService.checkEnclaveSignature(model, WALLET_ADDRESS)).isFalse();
+        assertThat(authorizationService.checkEnclaveSignature(model, workerCreds.getAddress())).isFalse();
     }
 
     @Test
     void shouldBeSignedByEnclave() {
-        final WorkerpoolAuthorization authorization = getWorkerpoolAuthorization();
+        final WorkerpoolAuthorization authorization = getWorkerpoolAuthorization(true);
         authorizationService.putIfAbsent(authorization);
         final ResultModel model = ResultModel.builder()
                 .chainTaskId(CHAIN_TASK_ID)
                 .enclaveSignature(getEnclaveSignature(enclaveCreds.getEcKeyPair()))
                 .deterministHash(RESULT_DIGEST)
                 .build();
-        assertThat(authorizationService.checkEnclaveSignature(model, WALLET_ADDRESS)).isTrue();
+        assertThat(authorizationService.checkEnclaveSignature(model, workerCreds.getAddress())).isTrue();
     }
     // endregion
 
-    // region
+    // region putIfAbsent
     @Test
     void shouldNotAddAuthorizationTwiceInCollection() {
-        final WorkerpoolAuthorization authorization = getWorkerpoolAuthorization();
-        authorizationService.putIfAbsent(authorization);
+        final WorkerpoolAuthorization stdAuthorization = getWorkerpoolAuthorization(false);
+        authorizationService.putIfAbsent(stdAuthorization);
         assertThat(authorizationRepository.count()).isOne();
-        authorizationService.putIfAbsent(authorization);
+        authorizationService.putIfAbsent(stdAuthorization);
+        assertThat(authorizationRepository.count()).isOne();
+        authorizationRepository.deleteAll();
+        final WorkerpoolAuthorization teeAuthorization = getWorkerpoolAuthorization(true);
+        authorizationService.putIfAbsent(teeAuthorization);
+        assertThat(authorizationRepository.count()).isOne();
+        authorizationService.putIfAbsent(teeAuthorization);
         assertThat(authorizationRepository.count()).isOne();
     }
     // endregion
 
     // region utils
-    ChainDeal getChainDeal() {
-        return ChainDeal.builder()
-                .poolOwner("0xc911f9345717ba7c8ec862ce002af3e058df84e4")
-                .tag(TeeUtils.TEE_SCONE_ONLY_TAG)
-                .build();
-    }
-
     String getEnclaveSignature(ECKeyPair ecKeyPair) {
         final String resultHash = HashUtils.concatenateAndHash(CHAIN_TASK_ID, RESULT_DIGEST);
-        final String resultSeal = HashUtils.concatenateAndHash(WALLET_ADDRESS, CHAIN_TASK_ID, RESULT_DIGEST);
+        final String resultSeal = HashUtils.concatenateAndHash(workerCreds.getAddress(), CHAIN_TASK_ID, RESULT_DIGEST);
         final String messageHash = HashUtils.concatenateAndHash(resultHash, resultSeal);
         return SignatureUtils.signMessageHashAndGetSignature(messageHash,
                 Numeric.toHexStringWithPrefix(ecKeyPair.getPrivateKey())).getValue();
     }
 
-    WorkerpoolAuthorization getWorkerpoolAuthorization() {
+    WorkerpoolAuthorization getWorkerpoolAuthorization(boolean isTeeTask) {
+        final String enclaveChallenge = isTeeTask ? enclaveCreds.getAddress() : BytesUtils.EMPTY_ADDRESS;
+        final String hash = HashUtils.concatenateAndHash(workerCreds.getAddress(), CHAIN_TASK_ID, enclaveChallenge);
+        final Signature signature = SignatureUtils.signMessageHashAndGetSignature(hash, POOL_PRIVATE);
         return WorkerpoolAuthorization.builder()
                 .chainTaskId(CHAIN_TASK_ID)
-                .enclaveChallenge(enclaveCreds.getAddress())
-                .workerWallet(WALLET_ADDRESS)
+                .enclaveChallenge(enclaveChallenge)
+                .workerWallet(workerCreds.getAddress())
+                .signature(signature)
                 .build();
     }
     // endregion